fastapi/requirements.txt
python-dotenv==1.0.1
fastapi==0.115.5
uvicorn[standard]==0.32.1
google-auth==2.36.0
google-api-python-client==2.154.0
google-auth-oauthlib==1.2.1
langchain==0.3.0
langchain-community==0.3.0
langchain-mistralai==0.2.0
langchain-qdrant==0.2.0
pypdf==5.1.0
unstructured==0.6.11
tabulate==0.9.0
pymupdf==1.25.1